Revert "[cxx-interop] Add tests for BorrowingForLoop"
This reverts commit 0367668b6265acb2c28e0567fe90ca8f807e2486.
