+++ test/Interop/Cxx/stdlib/overlay/custom-borrowing-sequence.swift
@@ -1,7 +1,6 @@
-// RUN: %target-run-simple-swift(-I %S/Inputs -Xfrontend -cxx-interoperability-mode=default -I %swift_src_root/lib/ClangImporter/SwiftBridging -Xcc -std=c++20 -enable-experimental-feature BorrowingForLoop)
+// RUN: %target-run-simple-swift(-I %S/Inputs -Xfrontend -cxx-interoperability-mode=default -I %swift_src_root/lib/ClangImporter/SwiftBridging -Xcc -std=c++20)
 //
 // REQUIRES: executable_test
-// REQUIRES: swift_feature_BorrowingForLoop
 // Ubuntu 20.04 ships with an old version of libstdc++, which does not provide
 // std::contiguous_iterator_tag from C++20.
 // UNSUPPORTED: LinuxDistribution=ubuntu-20.04
@@ -90,32 +89,6 @@ CxxBorrowingSequenceTestSuite.test("ContiguousNonCopyableSequence as Swift.Borro
   }
   expectEqual(innerCounter, 5)
   expectEqual(outerCounter, 2)
-}
-
-CxxBorrowingSequenceTestSuite.test("ContiguousNonCopyableSequence borrowing for loop") {
-  guard #available(SwiftStdlib 6.4, *) else { return }
-  let seq = ContiguousNonCopyableSequence()
-  let arr : [Int32] = [10, 20, 30, 40, 50]
-
-  var counter = 0
-  for el in seq {
-    expectEqual(el, arr[counter])
-    counter += 1
-  }
-  expectEqual(counter, 5)
-}
-
-CxxBorrowingSequenceTestSuite.test("DifferentResultsDereferenceOperatorSequence borrowing for loop") {
-  guard #available(SwiftStdlib 6.4, *) else { return }
-  let seq = DifferentResultsDereferenceOperatorSequence()
-  let arr : [Int32] = [2, 3, 4, 5]
-
-  var counter = 0
-  for el in seq {
-    expectEqual(el, arr[counter])
-    counter += 1
-  }
-  expectEqual(counter, 4)
 }
 
 runAllTests()
