+++ test/Interop/Cxx/stdlib/use-std-set.swift
@@ -1,19 +1,18 @@
-// RUN: %target-run-simple-swift(-I %S/Inputs -Xfrontend -enable-experimental-cxx-interop -enable-experimental-feature BorrowingForLoop)
-// RUN: %target-run-simple-swift(-I %S/Inputs -cxx-interoperability-mode=swift-6 -enable-experimental-feature BorrowingForLoop)
-// RUN: %target-run-simple-swift(-I %S/Inputs -cxx-interoperability-mode=upcoming-swift -enable-experimental-feature BorrowingForLoop)
-// RUN: %target-run-simple-swift(-I %S/Inputs -cxx-interoperability-mode=upcoming-swift -Xcc -std=c++14 -enable-experimental-feature BorrowingForLoop)
-// RUN: %target-run-simple-swift(-I %S/Inputs -cxx-interoperability-mode=upcoming-swift -Xcc -std=c++17 -enable-experimental-feature BorrowingForLoop)
-// RUN: %target-run-simple-swift(-I %S/Inputs -cxx-interoperability-mode=upcoming-swift -Xcc -std=c++20 -enable-experimental-feature BorrowingForLoop)
+// RUN: %target-run-simple-swift(-I %S/Inputs -Xfrontend -enable-experimental-cxx-interop)
+// RUN: %target-run-simple-swift(-I %S/Inputs -cxx-interoperability-mode=swift-6)
+// RUN: %target-run-simple-swift(-I %S/Inputs -cxx-interoperability-mode=upcoming-swift)
+// RUN: %target-run-simple-swift(-I %S/Inputs -cxx-interoperability-mode=upcoming-swift -Xcc -std=c++14)
+// RUN: %target-run-simple-swift(-I %S/Inputs -cxx-interoperability-mode=upcoming-swift -Xcc -std=c++17)
+// RUN: %target-run-simple-swift(-I %S/Inputs -cxx-interoperability-mode=upcoming-swift -Xcc -std=c++20)
 
 // Also test this with a bridging header instead of the StdSet module.
 // RUN: %empty-directory(%t2)
 // RUN: cp %S/Inputs/std-set.h %t2/std-set-bridging-header.h
-// RUN: %target-run-simple-swift(-D BRIDGING_HEADER -import-objc-header %t2/std-set-bridging-header.h -Xfrontend -enable-experimental-cxx-interop -enable-experimental-feature BorrowingForLoop)
-// RUN: %target-run-simple-swift(-D BRIDGING_HEADER -import-objc-header %t2/std-set-bridging-header.h -cxx-interoperability-mode=swift-6 -enable-experimental-feature BorrowingForLoop)
-// RUN: %target-run-simple-swift(-D BRIDGING_HEADER -import-objc-header %t2/std-set-bridging-header.h -cxx-interoperability-mode=upcoming-swift -enable-experimental-feature BorrowingForLoop)
+// RUN: %target-run-simple-swift(-D BRIDGING_HEADER -import-objc-header %t2/std-set-bridging-header.h -Xfrontend -enable-experimental-cxx-interop)
+// RUN: %target-run-simple-swift(-D BRIDGING_HEADER -import-objc-header %t2/std-set-bridging-header.h -cxx-interoperability-mode=swift-6)
+// RUN: %target-run-simple-swift(-D BRIDGING_HEADER -import-objc-header %t2/std-set-bridging-header.h -cxx-interoperability-mode=upcoming-swift)
 
 // REQUIRES: executable_test
-// REQUIRES: swift_feature_BorrowingForLoop
 //
 // Enable this everywhere once we have a solution for modularizing other C++ stdlibs: rdar://87654514
 // REQUIRES: OS=macosx || OS=linux-gnu
@@ -182,64 +181,6 @@ StdSetTestSuite.test("UnorderedSetOfCInt.remove") {
     expectFalse(s.contains(2))
     expectEqual(s.remove(2), nil)
     expectFalse(s.contains(2))
-}
-
-StdSetTestSuite.test("SetOfCInt Borrowing Iterators").require(.stdlib_6_4).code {
-    guard #available(SwiftStdlib 6.4, *) else { return }
-
-    let constSet = initSetOfCInt()
-    expectTrue(constSet.contains(1))
-    expectFalse(constSet.contains(2))
-    expectTrue(constSet.contains(3))
-
-    let arr : [Int32] = [1, 3, 5]
-    var counter = 0
-    for el in constSet {
-        expectEqual(el, arr[counter])
-        counter += 1
-    }
-    expectEqual(counter, 3)
-
-    var mutSet = initSetOfCInt()
-    expectTrue(mutSet.contains(1))
-    expectFalse(mutSet.contains(2))
-    expectTrue(mutSet.contains(3))
-
-    counter = 0
-    for el in mutSet {
-        expectEqual(el, arr[counter])
-        counter += 1
-    }
-    expectEqual(counter, 3)
-}
-
-StdSetTestSuite.test("MultisetOfCInt Borrowing Iterators").require(.stdlib_6_4).code {
-    guard #available(SwiftStdlib 6.4, *) else { return }
-
-    let constSet = initMultisetOfCInt()
-    expectTrue(constSet.contains(2))
-    expectFalse(constSet.contains(3))
-    expectTrue(constSet.contains(4))
-
-    let arr : [Int32] = [2, 2, 4, 6]
-    var counter = 0
-    for el in constSet {
-        expectEqual(el, arr[counter])
-        counter += 1
-    }
-    expectEqual(counter, 4)
-
-    var mutSet = initMultisetOfCInt()
-    expectTrue(mutSet.contains(2))
-    expectFalse(mutSet.contains(3))
-    expectTrue(mutSet.contains(4))
-
-    counter = 0
-    for el in mutSet {
-        expectEqual(el, arr[counter])
-        counter += 1
-    }
-    expectEqual(counter, 4)
 }
 
 runAllTests()
