+++ test/Interop/Cxx/stdlib/use-std-vector-typechecker.swift
@@ -1,20 +1,25 @@
-// RUN: %target-swift-frontend %s -typecheck -I %S/Inputs -cxx-interoperability-mode=default -verify
+// RUN: not %target-swift-frontend %s -typecheck -I %S/Inputs -cxx-interoperability-mode=default -diagnostic-style llvm 2>&1 | %FileCheck %s
 
 import StdVector
 import StdVectorNoCPlusPlusRequirement
 import CxxStdlib
 
-func takeCopyable<T: Copyable>(_ x: T) {} // expected-note *{{'where T: Copyable' is implicit here}}
-func takeCxxVector<V: CxxVector>(_ v: V) {} // expected-note {{where 'V' = 'VectorOfNonCopyable'}}
+func takeCopyable<T: Copyable>(_ x: T) {} 
+func takeCxxVector<V: CxxVector>(_ v: V) {} 
 
 let vecNC = VectorOfNonCopyable()
-takeCopyable(vecNC) // expected-error {{global function 'takeCopyable' requires that 'VectorOfNonCopyable'}}
+takeCopyable(vecNC)
+// CHECK: error: global function 'takeCopyable' requires that 'VectorOfNonCopyable' {{.*}} conform to 'Copyable'
+// CHECK: note: 'where T: Copyable' is implicit here
 
-takeCxxVector(vecNC) // expected-error {{global function 'takeCxxVector' requires that 'VectorOfNonCopyable'}}
+takeCxxVector(vecNC) 
+// CHECK: error: global function 'takeCxxVector' requires that 'VectorOfNonCopyable' {{.*}} conform to 'CxxVector'
+// CHECK: note: where 'V' = 'VectorOfNonCopyable' {{.*}}
 
 let vecPointer = VectorOfPointer()
 takeCopyable(vecPointer)
 takeCxxVector(vecPointer)
+// CHECK-NOT: error
 
 // Make sure that a specialization of std::vector that is declared in a Clang
 // module which does not declare 'requires cplusplus' is still conformed to
@@ -22,9 +27,12 @@ takeCxxVector(vecPointer)
 let vecFloat = VectorOfFloat()
 takeCopyable(vecFloat)
 takeCxxVector(vecFloat)
+// CHECK-NOT: error
 
 let hasVector = HasVector()
-takeCopyable(hasVector) // expected-error {{global function 'takeCopyable' requires that 'HasVector' conform to 'Copyable'}}
+takeCopyable(hasVector)
+// CHECK: error: global function 'takeCopyable' requires that 'HasVector' conform to 'Copyable'
 
 let baseHasVector = BaseHasVector()
-takeCopyable(baseHasVector) // expected-error {{global function 'takeCopyable' requires that 'BaseHasVector' conform to 'Copyable'}}
+takeCopyable(baseHasVector)
+// CHECK: error: global function 'takeCopyable' requires that 'BaseHasVector' conform to 'Copyable'
